backend/requirement.txt
aiosqlite==0.19.0
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.7.22
click==8.1.7
databases==0.8.0
dnspython==1.16.0
exceptiongroup==1.1.3
fastapi==0.104.1
greenlet==3.0.2
h11==0.14.0
httptools==0.6.1
idna==3.4
pydantic==2.5.0
pydantic_core==2.14.1
PyJWT==2.8.0
pymongo==4.3.3
PyMySQL==1.1.0
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==1.4.50
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
